README
This is a practice playground where I go to do coding challenges and try new methods in Python. This holds a library of methods I have used in the past that I like to reflect on whenever I am having an issue to see if I have encounter the same problem to some extent before.